refactor: move ShapesBackgroundAnimation to ramps directory
